longtail-producer/requirements.txt
Flask
kafka-python==2.0.1